feat: #66 모임 결과 조회 API 호출 함수 구현

+++ moyeo/moyeo/Model/MeetingListResponse.swift
@@ -10,7 +10,7 @@ import Foundation
 struct MeetingListResponse: Codable {
     
     struct MeetingStatus: Codable, Hashable {
-        let meetingId: Int64
+        let meetingId: Int
         let title: String
         let deadline: String
         let meetingStatus: String
